chore: move table and filter into separate component file
